fix(proc,logger): build on 32-bit GOARCH — close the regressions the #94 guard lanes caught (#94)
The linux/386 and linux/arm guard lanes added in 27da044 surfaced three real
32-bit portability bugs that only ever compiled on a 64-bit host:

- proc/exec/handle_unix.go: Rusage.Maxrss is int64 on a 64-bit GOARCH but int32
  on 386/arm — `return ru.Maxrss` does not compile where it is int32. Split the
  read into per-width files (maxrss_rss64_unix.go passes through, the int32 file
  widens to int64) so neither path carries a redundant cast on its own arch.
- proc/exec/cgroup_placement_linux.go: Statfs_t.Type is int32 on 386/arm, so the
  int64-typed cgroup2SuperMagic mismatched. Retype the magic as uint64 and
  compare uint64(st.Type) — a real (signed→unsigned) conversion on every GOARCH,
  never a redundant cast.
- logger/middleware/encwrite/encwrite.go: `len(box) > math.MaxUint32` overflows
  int on a 32-bit GOARCH (MaxUint32 does not fit int32). Widen len to uint64
  before the comparison.

All four modules now build on 386/arm/amd64/arm64; ktn-linter clean.

+++ internal/service/proc/exec/cgroup_placement_linux.go
@@ -24,9 +24,11 @@ import (
 const cgroupProcsFile string = "cgroup.procs"
 
 // cgroup2SuperMagic is the statfs f_type of the unified cgroup v2 filesystem
-// (CGROUP2_SUPER_MAGIC, the ASCII "cgrp"). Compared against int64(st.Type) so the
-// arch-dependent Statfs_t.Type field (int64 / uint32) widens cleanly.
-const cgroup2SuperMagic int64 = 0x63677270
+// (CGROUP2_SUPER_MAGIC, the ASCII "cgrp"). Typed uint64 and compared against
+// uint64(st.Type) so the arch-dependent Statfs_t.Type field (int64 on amd64/arm64,
+// int32 on 386/arm) widens cleanly — the conversion is real on every GOARCH (the
+// source is signed, the target unsigned), so it is never a redundant cast.
+const cgroup2SuperMagic uint64 = 0x63677270
 
 // cgroupProcsPerm is the perm passed to os.WriteFile; cgroup.procs always exists
 // so the mode is never used to create it — it mirrors the cgroup package's
@@ -63,7 +65,7 @@ func validateCgroupPath(path string) error {
 	}
 	//: the directory must sit on the cgroup2 filesystem — a normal directory that
 	//: merely contains a cgroup.procs file is NOT confinement and is rejected here.
-	if st.Type != cgroup2SuperMagic {
+	if uint64(st.Type) != cgroup2SuperMagic {
 		//: not a cgroup2 mount — refuse rather than spawn an unconfined child.
 		return wrapCgroupUnavailable(nil, errs.String("cgroup_path", path),
 			errs.String("reason", "path is not on a cgroup2 filesystem"))
